+++ packages/components/dialog/src/style.ts
@@ -4,11 +4,11 @@ export const test = "";
 export const dialog = tv({
   slots: {
     backdrop: "z-backdrop grid place-items-center backdrop-blur-sm",
-    header: "px-4 pt-4 text-base font-bold",
+    header: "px-2 pt-4 text-base font-bold md:px-4",
     body: "m-auto w-full flex-1 overflow-y-auto p-2 md:p-4",
     content:
-      "relative z-dialog flex  w-full flex-col rounded-3xl bg-background-secondary shadow-lg backdrop-blur-3xl",
-    footer: "bottom-0 flex w-full justify-end p-2",
+      "relative z-dialog flex  w-full flex-col rounded-3xl bg-content1 shadow-lg backdrop-blur-3xl",
+    footer: "bottom-0 flex w-full justify-end gap-2 p-2 md:p-4",
   },
   variants: {
     size: {
@@ -54,7 +54,7 @@ export const dialog = tv({
         backdrop: "bg-overlay/30 backdrop-blur-md backdrop-saturate-150",
       },
     },
-    scroll: {
+    isScroll: {
       true: {
         header: "border-b border-divider pb-4",
         footer: "sticky border-t border-divider",