+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/accordion
 
+## 0.0.73
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+
 ## 0.0.72
 
 ### Patch Changes