+++ packages/components/alert/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/alert
 
+## 0.0.80
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+  - @jamsr-ui/typography@0.0.71
+
 ## 0.0.79
 
 ### Patch Changes