+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/card
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+  - @jamsr-ui/typography@0.0.71
+
 ## 0.0.75
 
 ### Patch Changes