+++ packages/components/confirmation/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsr-ui/confirmation
 
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/divider@0.0.64
+  - @jamsr-ui/dialog@0.0.73
+  - @jamsr-ui/utils@0.0.60
+  - @jamsr-ui/button@0.0.66
+  - @jamsr-ui/typography@0.0.71
+
 ## 0.0.49
 
 ### Patch Changes