+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/dialog
 
+## 0.0.73
+
+### Patch Changes
+
+- fix packages
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+  - @jamsr-ui/button@0.0.66
+
 ## 0.0.72
 
 ### Patch Changes