+++ packages/components/sidebar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/sidebar
 
+## 0.0.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+
 ## 0.0.71
 
 ### Patch Changes