+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/switch
 
+## 0.0.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+  - @jamsr-ui/typography@0.0.71
+
 ## 0.0.71
 
 ### Patch Changes