+++ packages/components/table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/table
 
+## 0.0.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.60
+
 ## 0.0.42
 
 ### Patch Changes