+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/utils
 
+## 0.0.60
+
+### Patch Changes
+
+- fix packages
+
 ## 0.0.59
 
 ### Patch Changes